add queues optimizations
- The way this optimization works is by creating a queue for each source
of messages.

- When we receive a message from a source we can safely insert that
message in the its queue. All the messages from this queue will be
totally ordered, since we cannot receive, from the same source, two
messages that are concurrent.

- With this system it's only needed to check if only the first message
from each list is possible to be delivered, since the queues are
ordered. However if one queue is able to deliver at least one message we
need to check all the queues again. This happens because by delivering a
message we might be enabling another one to be delivered. That said when
all the queues don't have a message ready to be delivered we can safely
say that we can stop searching for messages to be delivered.