fix credits increment
- The goal should had been to allow up to n messages in the quarantine
queue and delivered queue. When we insert to a new queue we decrement
the available credits. When a message dequeues from the delivered queue
we increment our available credits and call request to parent. This
means we will have stored at any time at most n messages in our memory.